requirements.txt
torch==2.0.1
torchvision==0.15.2
kornia==0.6.5
opencv-python==4.6.0.66
matplotlib==3.7.1
PyYAML==6.0
scikit-learn==1.2.2
Pillow==10.0.0